+++ gnmatcher.go
@@ -7,6 +7,7 @@ import (
 	"github.com/gnames/gnmatcher/ent/exact"
 	"github.com/gnames/gnmatcher/ent/fuzzy"
 	"github.com/gnames/gnmatcher/ent/matcher"
+	"github.com/gnames/gnmatcher/ent/virus"
 )
 
 // gnmatcher implements GNmatcher interface.
@@ -17,9 +18,14 @@ type gnmatcher struct {
 
 // New is a constructor for GNmatcher interface. It takes two
 // interfaces ExactMatcher and FuzzyMatcher.
-func New(em exact.ExactMatcher, fm fuzzy.FuzzyMatcher, cfg config.Config) GNmatcher {
+func New(
+	em exact.ExactMatcher,
+	fm fuzzy.FuzzyMatcher,
+	vm virus.VirusMatcher,
+	cfg config.Config,
+) GNmatcher {
 	gnm := gnmatcher{cfg: cfg}
-	gnm.matcher = matcher.NewMatcher(em, fm, cfg.JobsNum)
+	gnm.matcher = matcher.NewMatcher(em, fm, vm, cfg.JobsNum)
 	gnm.matcher.Init()
 	return gnm
 }